+++ test/Finbuckle.MultiTenant.AspNetCore.Test/Strategies/SessionStrategyShould.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Threading.Tasks;
+using Finbuckle.MultiTenant.DependencyInjection;
 using Finbuckle.MultiTenant.Strategies;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
